Correções no serviço de algoritmos

+++ src/ProgramAcad.Domain/Entities/ExecucaoTeste.cs
@@ -5,7 +5,7 @@ namespace ProgramAcad.Domain.Entities
     public class ExecucaoTeste
     {
         public ExecucaoTeste(Guid idCasoTeste, Guid idUsuario, LinguagensProgramacao idLinguagem, 
-            bool sucesso, int tempoExecucao)
+            bool sucesso, double tempoExecucao)
         {
             Id = Guid.NewGuid();
             IdCasoTeste = idCasoTeste;
@@ -20,7 +20,7 @@ namespace ProgramAcad.Domain.Entities
         public Guid IdUsuario { get; protected set; }
         public LinguagensProgramacao IdLinguagem { get; protected set; }
         public bool Sucesso { get; protected set; }
-        public int TempoExecucao { get; protected set; }
+        public double TempoExecucao { get; protected set; }
 
 
         public CasoTeste CasoTeste { get; set; }
