+++ src/ProgramAcad.Services/Modules/Compiling/Models/CompilerResponse.cs
@@ -4,7 +4,7 @@ namespace ProgramAcad.Services.Modules.Compiling.Models
 {
     public class CompilerResponse
     {
-        public CompilerResponse(string output, HttpStatusCode statusCode, double? cpuTime)
+        public CompilerResponse(string output, HttpStatusCode statusCode, float? cpuTime)
         {
             Output = output;
             StatusCode = statusCode;
@@ -13,7 +13,7 @@ namespace ProgramAcad.Services.Modules.Compiling.Models
 
         public string Output { get; protected set; }
         public HttpStatusCode StatusCode { get; protected set; }
-        public double? CpuTime { get; protected set; }
+        public float? CpuTime { get; protected set; }
 
         public bool HasCompilingError => CpuTime == null || (Output != null && Output.Contains("Traceback (most recent call last)"));
     }
